 requirements.txt
protobuf==4.25.1
numpy==1.26.3
loguru==0.7.2
playwright==1.41.0
requests ==2.31.0
torch==2.2.0